Move asset retrieval to kibana backend
