+++ server/adaptors/placeholder/__tests__/test.ndjson
