+++ server/routes/placeholder/placeholder_router.ts
@@ -5,12 +5,7 @@
 
 import fetch from 'node-fetch';
 import * as fs from 'fs';
-import { Readable } from 'stream';
-import {
-  ILegacyScopedClusterClient,
-  IRouter,
-  RequestHandlerContext,
-} from '../../../../../src/core/server';
+import { IRouter, RequestHandlerContext } from '../../../../../src/core/server';
 import { INTEGRATIONS_BASE, OBSERVABILITY_BASE } from '../../../common/constants/shared';
 import { PlaceholderAdaptor } from '../../../server/adaptors/placeholder/placeholder_adaptor';
 import {
@@ -19,39 +14,6 @@ import {
 } from '../../../../../src/core/server/http/router';
 import { SavedObjectsBulkCreateObject } from '../../../../../src/core/public';
 import { PlaceholderKibanaBackend } from '../../../server/adaptors/placeholder/placeholder_kibana_backend';
-
-/**
- * Parse a stream of newline-delimited JSON objects as an array.
- * The entire stream is read; the method will hang if the stream is not closed.
- * The data entries MUST be JSON objects,
- * other valid JSON values will be rejected.
- *
- * Resolves the `Promise` if every newline-separated JSON object is valid.
- * Rejects the `Promise` if the stream errors, or if the JSON is not parseable.
- *
- * @param {Readable} stream A stream of newline-delimited JSON objects.
- * @returns {Promise<object[]>} A `Promise` for an array of parsed JSON objects.
- */
-export const readNDJsonObjects = async (stream: Readable): Promise<object[]> => {
-  return new Promise<object[]>((resolve, reject) => {
-    let assets: object[] = [];
-    let json: string = '';
-    stream.on('data', (chunk: string | Buffer) => {
-      json += chunk.toString();
-    });
-    stream.on('end', () => {
-      try {
-        assets = JSON.parse(`[${json.replace(/\}\s+\{/g, '},{')}]`);
-        resolve(assets);
-      } catch (err: any) {
-        reject(err);
-      }
-    });
-    stream.on('error', (err: Error) => {
-      reject(err);
-    });
-  });
-};
 
 let added = false;
 
@@ -119,9 +81,8 @@ export function registerPlaceholderRoute(router: IRouter) {
     },
     async (context, request, response): Promise<any> => {
       const adaptor = getAdaptor(context, request);
-      return handleWithCallback(adaptor, response, async (_a: PlaceholderAdaptor) => {
-        const stream = fs.createReadStream(__dirname + '/__tests__/test.ndjson');
-        const assets = (await readNDJsonObjects(stream)) as SavedObjectsBulkCreateObject[];
+      return handleWithCallback(adaptor, response, async (a: PlaceholderAdaptor) => {
+        const assets = await a.getAssets('nginx');
         added = true;
         return context.core.savedObjects.client.bulkCreate(assets);
       });
